requirements/base.txt
asgiref==3.8.1

# Django
Django==4.1
django-extensions==3.2.3
django-cors-headers==4.5.0
django-storages==1.14.3

# env
django-environ==0.11.2
python-dotenv==0.21.1

# DRF
djangorestframework==3.14.0

# JWT
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0

pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2

# Filters
django-filter==23.5

# Swagger
drf-yasg==1.21.7
setuptools==75.1.0

# Pytest
pytest-django==4.9.0

# Images
pillow==11.0.0

# postgresql
psycopg2==2.9.10
dj-database-url==2.1.0

# Reports
pandas==2.2.3
django-pandas==0.6.7
XlsxWriter==3.0.2
